blog detail page added

+++ BlogAppCore/Core/IRepositories/IGenericRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IGenericRepository<T> where T : class
     {
-        public Task<T> GetByIdAsync(int id);
+        public Task<T> GetByIdAsync(long id);
         public Task<IEnumerable<T>> GetAllAsync();
         public Task<bool> AddAsync(T entity);
         public Task<bool> UpdateAsync(T entity);
